requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
streamlit==1.28.0
langchain==0.0.335
openai==1.2.4
python-dotenv==1.0.0
pydantic==1.10.13
requests==2.31.0